updated CanCardDetails component
