+++ frontend/src/pages/portfolios/detail/PortfolioDetail.jsx
@@ -1,4 +1,4 @@
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { getPortfolioAndSetState } from "./getPortfolio";
 import { useEffect } from "react";
 import { useParams } from "react-router-dom";
@@ -26,29 +26,29 @@ const PortfolioDetail = () => {
     const portfolioId = parseInt(urlPathParams.id);
     const canId = 1;
     const currentFiscalYear = getCurrentFiscalYear(new Date());
-    //const portfolioCans = useSelector((state) => state.portfolioDetail.portfolioCans);
-    const portfolioCans = [
-        {
-            arrangement_type_id: 3,
-            authorizer_id: 26,
-            description: "Incoming Interagency Agreements",
-            id: 2,
-            managing_portfolio_id: 1,
-            nickname: "IAA-Incoming",
-            number: "G99IA14",
-            purpose: null,
-        },
-        {
-            arrangement_type_id: 4,
-            authorizer_id: 26,
-            description: "Child Development Research Fellowship Grant Program",
-            id: 4,
-            managing_portfolio_id: 1,
-            nickname: "ASPE SRCD-IDDA",
-            number: "G990136",
-            purpose: null,
-        },
-    ];
+    const portfolioCans = useSelector((state) => state.portfolioDetail.portfolioCans);
+    // const portfolioCans = [
+    //     {
+    //         arrangement_type_id: 3,
+    //         authorizer_id: 26,
+    //         description: "Incoming Interagency Agreements",
+    //         id: 2,
+    //         managing_portfolio_id: 1,
+    //         nickname: "IAA-Incoming",
+    //         number: "G99IA14",
+    //         purpose: null,
+    //     },
+    //     {
+    //         arrangement_type_id: 4,
+    //         authorizer_id: 26,
+    //         description: "Child Development Research Fellowship Grant Program",
+    //         id: 4,
+    //         managing_portfolio_id: 1,
+    //         nickname: "ASPE SRCD-IDDA",
+    //         number: "G990136",
+    //         purpose: null,
+    //     },
+    // ];
     useEffect(() => {
         dispatch(getPortfolioAndSetState(portfolioId));
         return () => {
